Rename: Track 엔티티 위치 이동 (track -> track.entity)
#9
